backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
motor==3.3.2
pymongo==4.6.0

# Configuration and utilities
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic[email]==2.5.0

# Authentication and Security
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
email-validator==2.1.0
cryptography==41.0.1
aiofiles==23.2.1
boto3==1.29.6

# AI and LLM integration
langchain>=0.0.350
langchain-openai>=0.0.2.post1
openai>=1.6.1
tiktoken>=0.5.1
pika>=1.3.2

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1